requirements.txt
pymupdf>=1.24.0
torch>=2.1.0
transformers>=4.45.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
streamlit>=1.28.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
langfuse>=2.0.0
pytest>=7.4.0
pandas>=2.0.0
ragas>=0.2.0
rank-bm25>=0.2.2
scikit-learn>=1.3.0
pdfplumber>=0.11.0
Pillow>=10.0.0
langchain-text-splitters>=0.2.0
